requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.6.1
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
Markdown==3.2.2
mysqlclient==1.4.6
numpy==1.19.0
pandas==1.0.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.4
zipp==3.1.0
